# lumio-be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
    "/cart/items/{courseId}": {
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/webhooks/stripe": {
      "post": {
        "operationId": "StripeWebhookController_handle",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/orders/{orderId}/refund": {
      "post": {
        "operationId": "RefundsController_request",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/refunds/{id}/approve": {
      "post": {
        "operationId": "RefundsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
    "/refunds/{id}/reject": {
      "post": {
        "operationId": "RefundsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/instructor/courses/{id}/analytics": {
      "get": {
        "operationId": "InstructorController_analytics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "instructor"
        ]
      }
    },
...
    "/courses/{id}/announcements": {
      "post": {
        "operationId": "InstructorController_createAnnouncement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "instructor"
        ]
      },
      "get": {
        "operationId": "InstructorController_listAnnouncements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "instructor"
        ]
      }
    },
...
    "/org-admin/users/{userId}/suspend": {
      "post": {
        "operationId": "OrgAdminController_suspend",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "org-admin"
        ]
      }
    },
    "/org-admin/users/{userId}/reinstate": {
      "post": {
        "operationId": "OrgAdminController_reinstate",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "org-admin"
        ]
      }
    },
...
    "/super-admin/plans": {
      "post": {
        "operationId": "PlatformBillingController_createPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "platform-billing"
        ]
      }
    },
    "/tenants/me/subscription": {
      "get": {
        "operationId": "PlatformBillingController_getMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "platform-billing"
        ]
      },
      "post": {
        "operationId": "PlatformBillingController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "platform-billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumio LMS API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddCartItemDto": {
        "type": "object",
        "properties": {}
      },
...
      "RequestRefundDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAnnouncementDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "STARTER",
              "GROWTH",
              "ENTERPRISE"
            ]
          }
        },
        "required": [
          "code"
        ]
      },
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "enum": [
              "STARTER",
              "GROWTH",
              "ENTERPRISE"
            ]
          }
        },
        "required": [
          "planCode"
        ]
      },
...
    }
  }
}